backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
PyJWT==2.10.1
werkzeug==3.0.1

